+++ packages/extension-logger/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.3](https://github.com/ueberdosis/hocuspocus/compare/v3.2.2...v3.2.3) (2025-08-06)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
+## [3.2.2](https://github.com/ueberdosis/hocuspocus/compare/v3.2.1...v3.2.2) (2025-07-28)
+
+**Note:** Version bump only for package @hocuspocus/extension-logger
+
+
+
+
+
 ## [3.2.1](https://github.com/ueberdosis/hocuspocus/compare/v3.2.0...v3.2.1) (2025-07-20)
 
 **Note:** Version bump only for package @hocuspocus/extension-logger